+++ mowl/nn/__init__.py
@@ -1,4 +1,5 @@
 from .el.elmodule import ELModule
 from .el.elem.module import ELEmModule
-from .el.elbox.module import ELBoxModule
+from .el.elbe.module import ELBoxModule, ELBEModule
+from .el.boxel.module import BoxELModule
 from .el.boxsquaredel.module import BoxSquaredELModule