+++ mowl/nn/el/elem/module.py
@@ -2,8 +2,9 @@ import mowl.nn.el.elem.losses as L
 from mowl.nn import ELModule
 from torch import nn
 import torch as th
+from deprecated.sphinx import versionchanged
 
-
+@versionchanged(version="0.4.0", reason="The class ELEmModule receives an optional parameter nb_inds")
 class ELEmModule(ELModule):
     """
     Implementation of ELEmbeddings from [kulmanov2019]_.
@@ -11,10 +12,11 @@ class ELEmModule(ELModule):
     """
     
     
-    def __init__(self, nb_ont_classes, nb_rels, embed_dim=50, margin=0.1, reg_norm=1):
+    def __init__(self, nb_ont_classes, nb_rels, nb_inds=None, embed_dim=50, margin=0.1, reg_norm=1):
         super().__init__()
         self.nb_ont_classes = nb_ont_classes
         self.nb_rels = nb_rels
+        self.nb_inds = nb_inds
         self.reg_norm = reg_norm
         self.embed_dim = embed_dim
 
@@ -37,36 +39,57 @@ class ELEmModule(ELModule):
         weight_data_normalized = th.linalg.norm(self.rel_embed.weight.data, axis=1).reshape(-1, 1)
         self.rel_embed.weight.data /= weight_data_normalized
 
+        if self.nb_inds is not None:
+            self.ind_embed = nn.Embedding(self.nb_inds, embed_dim)
+            nn.init.uniform_(self.ind_embed.weight, a=-1, b=1)
+            weight_data_normalized = th.linalg.norm(self.ind_embed.weight.data, axis=1).reshape(-1, 1)
+            self.ind_embed.weight.data /= weight_data_normalized
+        else:
+            self.ind_embed = None
+            
+
         self.margin = margin
 
-    def class_reg(self, x):
-        # force n-ball to be inside unit ball
-        res = th.abs(th.linalg.norm(x, axis=1) - self.reg_norm)
-        res = th.reshape(res, [-1, 1])
-        return res
-
     def gci0_loss(self, data, neg=False):
-        return L.gci0_loss(data, self.class_embed, self.class_rad, self.class_reg, self.margin,
+        return L.gci0_loss(data, self.class_embed, self.class_rad, self.margin,
                            neg=neg)
 
     def gci0_bot_loss(self, data, neg=False):
         return L.gci0_bot_loss(data, self.class_rad)
         
     def gci1_loss(self, data, neg=False):
-        return L.gci1_loss(data, self.class_embed, self.class_rad, self.class_reg, self.margin,
+        return L.gci1_loss(data, self.class_embed, self.class_rad, self.margin,
                            neg=neg)
 
     def gci1_bot_loss(self, data, neg=False):
-        return L.gci1_bot_loss(data, self.class_embed, self.class_rad, self.class_reg, self.margin,
+        return L.gci1_bot_loss(data, self.class_embed, self.class_rad, self.margin,
                                neg=neg)
 
     def gci2_loss(self, data, neg=False, idxs_for_negs=None):
-        return L.gci2_loss(data, self.class_embed, self.class_rad, self.rel_embed, self.class_reg,
+        return L.gci2_loss(data, self.class_embed, self.class_rad, self.rel_embed,
                            self.margin, neg=neg)
 
     def gci3_loss(self, data, neg=False):
-        return L.gci3_loss(data, self.class_embed, self.class_rad, self.rel_embed, self.class_reg,
+        return L.gci3_loss(data, self.class_embed, self.class_rad, self.rel_embed,
                            self.margin, neg=neg)
 
     def gci3_bot_loss(self, data, neg=False):
         return L.gci3_bot_loss(data, self.class_rad)
+
+
+    def gci2_score(self, data):
+        return L.gci2_score(data, self.class_embed, self.class_rad, self.rel_embed, self.margin)
+
+    def class_assertion_loss(self, data, neg=False):
+        if self.ind_embed is None:
+            raise ValueError("The number of individuals must be specified to use this loss function.")
+        return L.class_assertion_loss(data, self.class_embed, self.class_rad, self.ind_embed, self.margin, neg=neg)
+
+    def object_property_assertion_loss(self, data, neg=False):
+        if self.ind_embed is None:
+            raise ValueError("The number of individuals must be specified to use this loss function.")
+        return L.object_property_assertion_loss(data, self.rel_embed, self.ind_embed, self.margin, neg=neg)
+
+    
+    def regularization_loss(self):
+        return L.regularization_loss(self.class_embed, self.reg_norm)