+++ CHANGELOG.md
@@ -7,10 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Added `load_corpus` to `mowl.base_models.SyntacticModel`
+- Added `.jar` files to enable pip install from GitHub
+- Tested with Python 3.12
 ### Changed 
+- Removed `.mean()` for GCI losses in BoxSquaredELModule
+- Property `evaluation_classes` in `mowl.datasets.builtin.PPIYeastDataset` returns a pair of `OWLClasses` objects instead of a single `OWLClasses` object.
 ### Deprecated
+- `mowl.nn.ELBoxModule` changed name to `mowl.nn.ELBEModule`
 ### Removed
 ### Fixed
+- Fix bug in GCI2 score for ELEmbeddings
+- Fix bottleneck in ELBE example for PPI.
+- Fix bugs in BoxSquaredEL model.
+
 ### Security
 
 ## [0.3.0]